modularize repo by creating new subdirectories
