+++ shellhelperfiles/shellhelpers1.c
@@ -1,4 +1,4 @@
-#include "shell.h"
+#include "../shell.h"
 
 /**
  * sig_handler - handles interruption signal
