+++ shellhelperfiles/shellhelpers2.c
@@ -1,4 +1,4 @@
-#include "shell.h"
+#include "../shell.h"
 
 /**
  * _getenv - get environment variable
